+++ internal/user/repo.go
@@ -129,3 +129,35 @@ func (r *Repo) GetLastActivityInPeriod(userID uuid.UUID, window time.Duration) (
 
 	return &activity, nil
 }
+
+func (r *Repo) GetLastActivity(userID uuid.UUID) (*Activity, error) {
+	var activity Activity
+	req := r.db.
+		Model(&Activity{}).
+		Where("user_id = ?", userID).
+		Order("finished_at desc").
+		First(&activity)
+
+	if err := req.Error; err != nil {
+		return nil, err
+	}
+
+	return &activity, nil
+}
+
+func (r *Repo) GetByFilters(filters []Filter) ([]Activity, error) {
+	db := r.db
+	for _, f := range filters {
+		if _, ok := f.(PageFilter); ok {
+			db = f.Apply(db)
+		}
+	}
+
+	var list []Activity
+	err := db.Find(&list).Error
+	if err != nil {
+		return nil, err
+	}
+
+	return list, nil
+}