+++ CHANGELOG.md
@@ -10,6 +10,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [0.0.15] - 2024-01-31
 
 ### Added
+- Calculating intervals for sending push notifications
+
+### Added
 - Tracking user activity endpoint
 - Added user nonce usage for replay protection